chg: dev: Sidelining old unit and integration tests, transitioning to tests and init structure for new sensor decomposer/correlator internals. !minor
