+++ sitch/sitchlib/__init__.py
@@ -1,10 +1,10 @@
 from alert_manager import AlertManager  # NOQA
 from config_helper import ConfigHelper  # NOQA
-from enricher import Enricher  # NOQA
+# from enricher import Enricher  # NOQA
 from gps_device import GpsListener  # NOQA
 from gsm_modem import GsmModem  # NOQA
-from enrich_gsm_modem import GsmModemEnricher  # NOQA
-from enrich_arfcn import EnrichArfcn  # NOQA
+# from enrich_gsm_modem import GsmModemEnricher  # NOQA
+# from enrich_arfcn import EnrichArfcn  # NOQA
 from location_tool import LocationTool  # NOQA
 from logger import LogHandler  # NOQA
 from geo_ip import GeoIp  # NOQA
@@ -12,14 +12,15 @@ from gsm_modem import GsmModem  # NOQA
 from utility import Utility  # NOQA
 from feed_manager import FeedManager  # NOQA
 from device_detector import DeviceDetector  # NOQA
-from enrich_kal_scan import KalScanEnricher  # NOQA
+#from enrich_kal_scan import KalScanEnricher  # NOQA
 
-from gps_decomposer import GpsDecomposer  # NOQA
-from gsm_decomposer import GsmDecomposer  # NOQA
-from kal_decomposer import KalDecomposer  # NOQA
+from decomposer import Decomposer  # NOQA
+# from gps_decomposer import GpsDecomposer  # NOQA
+# from gsm_decomposer import GsmDecomposer  # NOQA
+# from kal_decomposer import KalDecomposer  # NOQA
 
-from correlate_arfcn import ArfcnCorrelator  # NOQA
-from correlate_cgi import CgiCorrelator  # NOQA
+from arfcn_correlator import ArfcnCorrelator  # NOQA
+from cgi_correlator import CgiCorrelator  # NOQA
 
 __author__ = "Ash Wilson"
 __version__ = "3.5.0"
